try to benchmark an aligned adler32 variant
it just has the same performance/instructions/branches though
